Add tests for field values being too large
The spec gives a limit of 32-bits for the values of VLQ fields.
However, it's not very precise on how this limit is implemented.
For example, it's not clear if the sign bit is included or not.

For these tests, we have assumed for now that values exceeding
32-bits are invalid and makes the mapping invalid. In addition,
we have assumed the sign bit should be included in the limit.