Refine testing for sources/sourcesContent fields
It's not explicitly spelled out in the spec (FIXME) but it doesn't make sense
to have a null source and source content for a particular source entry, so it
might make sense to make that invalid. It should likely be valid to have a null
source entry as long as a source content exists.