Change behavior of basic-mapping-as-index-map test
It looks like this approach of having the same source appear in multiple
sections doesn't really match with how tools process source maps. It would
be better to have a multi-section test with concatenated sources. The
spec should probably also be clear on whether, e.g., the sources lists
should be mutually exclusive.