Change both null sources/sourcesContent case to valid
While it initially seems like this case should be invalid, I think there is a
good argument to be made that it's valid. First, the spec doesn't explicitly
rule it out and does allow the possibility of both being null.

Second, in practice it does have a meaning for browser engines. For example,
if you have a source map with null as a source, and also have a file named
"null" present then the browser may load that as the reference source file.